buildfix for some wiimote plugin errors
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@4651 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Plugins/Plugin_Wiimote/Src/Rumble.cpp
@@ -228,7 +228,7 @@ void SetDeviceForcesXY(int npad, int nXYForce)
 	{
 		rglDirection[0] = nXYForce;
 		rglDirection[1] = nXYForce;
-		cf.lMagnitude = 1.4142f*nXYForce;
+		cf.lMagnitude = static_cast<LONG>(1.4142f*nXYForce);
 	}
 
 	ZeroMemory(&pRumble[npad].eff, sizeof(pRumble[npad].eff));
